feat: noteId field in NotesSchema has index now, change getNotes in NotesController, some fixes in other methods this controller

+++ controllers/notes.js
@@ -1,34 +1,31 @@
 import Note from '../database/models/notes.js'
+import Page from '../database/models/pages.js'
 
 class NotesController {
   async getNotes(req, rep) {
-    return await Note.find({ author: req.currentUser.userId })
+    const foundPage = await Page.findOne({ _id: req.params.pageId })
+    if(foundPage.author !== req.currentUser.userId) throw new Error('You don\'t have permission')
+
+    return await Note.find({ page: req.params.pageId })
   }
 
   async createMultipleNotes(req, rep) {
-    const notesToAdd = req.body.notesToAdd.map((note) => {
-      note.author = req.currentUser.userId
-    })
+    const notesToAdd = req.body.notesToAdd.map((note) => ({
+      ...note, author: req.currentUser.userId
+    }))
     await Note.insertMany(notesToAdd)
-    return { addedNotes: req.body.notesToAdd.length }
+    return { success: true }
   }
 
   async removeMultipleNotes(req, rep) {
-    req.body.notesToRemove.forEach((note) => {
-      if(note.author !== req.currentUser.userId) throw new Error('You don\'t have permissions')
-    })
-    await Note.deleteMany({ noteId: req.body.notesToRemove })
-    return { removedNotes: req.body.notesToRemove.length }
+    await Note.deleteMany({ noteId: req.body.notesToRemove, author: req.currentUser.userId })
+    return { success: true }
   }
 
   async updateMultipleNotes(req, rep) {
-    req.body.changedNotes.forEach((note) => {
-      if(note.author !== req.currentUser.userId) throw new Error('You don\'t have permissions')
-    })
-
     const operations = req.body.changedNotes.map((note) => ({
       updateOne: {
-        filter: { noteId: note.noteId }, update: { $set: note }
+        filter: { noteId: note.noteId, author: req.currentUser.userId }, update: { $set: note }
       }
     }))
     await Note.bulkWrite(operations)
